move checkPermission to a standalone file and remove unused code
