+++ backend/pkg/casbin/casbin.go
@@ -102,13 +102,3 @@ func InitializeCasbin(pbClient *pocketbase.PocketBaseClient) (*CasbinEnforcer, e
 	fmt.Println("Casbin Enforcer initialized successfully.")
 	return &CasbinEnforcer{Enforcer: e}, nil
 }
-
-// CheckPermission validates user actions using Casbin
-func (ce *CasbinEnforcer) CheckPermission(roleId, resource, action, scope string) bool {
-	ok, err := ce.Enforcer.Enforce(roleId, resource, action, scope)
-	if err != nil {
-		fmt.Println("Error enforcing policy:", err)
-		return false
-	}
-	return ok
-}
